app/requirements.txt
python-dotenv
streamlit
requests
diskcache
plotly
extra-streamlit-components